Add UnrealFinderTool config file for DQXIS + minor code fix
Not 100% sure how accurate the defs inside are, but I do know that other 4.18 games don't require these changes, so they must be customizations added by DQXIS devs

+++ DQXI-SDK/FirstPerson.cpp
@@ -193,7 +193,7 @@ void* CameraInterpolate_Hook(UObject* a1, void* a2, void* a3)
   if (!NumFramesToZeroInterpSpeed)
     return CameraInterpolate_Orig(a1, a2, a3);
 
-  for (int i = 0; i < CameraDatas.size(); i++)
+  for (size_t i = 0; i < CameraDatas.size(); i++)
   {
     CameraDatas[i]->InterpTargetLocation_Speed = 0;
     CameraDatas[i]->InterpTargetLocation_SpeedXY = 0;
@@ -205,7 +205,7 @@ void* CameraInterpolate_Hook(UObject* a1, void* a2, void* a3)
 
   auto ret = CameraInterpolate_Orig(a1, a2, a3);
 
-  for (int i = 0; i < CameraDatas.size(); i++)
+  for (size_t i = 0; i < CameraDatas.size(); i++)
   {
     CameraDatas[i]->InterpTargetLocation_Speed = OriginalCameraDatas[i].InterpTargetLocation_Speed;
     CameraDatas[i]->InterpTargetLocation_SpeedXY = OriginalCameraDatas[i].InterpTargetLocation_SpeedXY;
